Reworked footer
While working with the footer I realized that its function is basically limited to the lessons page, so I decided to simply embed it in lessons.html instead of importing it. I also decided that it'd be better to just have next and last link to adjacent lessons since that's more intuitive.